+++ ser/transform.py
@@ -1,0 +1,11 @@
+import torch
+from torchvision import transforms
+
+
+def get_transform():
+    # torch transforms
+    ts = transforms.Compose(
+        [transforms.ToTensor(), transforms.Normalize((0.5,), (0.5,))]
+    )
+
+    return ts
